fix: 필요한 생성자 추가, section 관련 클래스 분리
