+++ src/main/java/nextstep/subway/line/ui/LineController.java
@@ -50,10 +50,4 @@ public class LineController {
         return ResponseEntity.noContent().build();
     }
 
-    @PutMapping("/lines/{lineId}/sections")
-    public ResponseEntity addSection(@PathVariable Long lineId,
-                                     @RequestBody SectionRequest sectionRequest) {
-        lineService.addSection(lineId, sectionRequest);
-        return ResponseEntity.ok().build();
-    }
 }
