+++ src/main/java/nextstep/subway/section/ui/SectionController.java
@@ -1,0 +1,21 @@
+package nextstep.subway.section.ui;
+
+import nextstep.subway.section.dto.SectionRequest;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+public class SectionController {
+    private final SectionService sectionService;
+
+    public SectionController(SectionService sectionService) {
+        this.sectionService = sectionService;
+    }
+
+    @PostMapping("/lines/{lineId}/sections")
+    public ResponseEntity addSection(@PathVariable Long lineId,
+                                     @RequestBody SectionRequest sectionRequest) {
+        sectionService.addSection(lineId, sectionRequest);
+        return ResponseEntity.ok().build();
+    }
+}
