+++ src/main/resources/application.properties
@@ -3,3 +3,5 @@ handlebars.enabled=true
 
 spring.jpa.properties.hibernate.show_sql=true
 spring.jpa.properties.hibernate.format_sql=true
+
+logging.level.root=DEBUG
